Move lodepng to external.
