+++ external/lodepng/lodepng.cpp
