+++ tools/ktx/command_extract.cpp
@@ -23,7 +23,7 @@
 #include <fmt/ostream.h>
 #include <fmt/printf.h>
 
-#include "png.imageio/lodepng.h"
+#include "lodepng/lodepng.h"
 #include "astc-encoder/Source/ThirdParty/tinyexr.h"
 #include "astc-encoder/Source/astcenc.h"
